+++ tests/testCaballo.py
@@ -26,8 +26,10 @@ class TestKnight(unittest.TestCase):
         knight = Knight("BLACK", board)
         self.assertEqual(knight.symbol(), '♞')
 
-# Testeo movimientos válidos
+# TESTEO MOVIMIENTOS
 
+
+    # Testeo de movimientos válidos
     def test_valid_moves(self):
         valid_moves = [
             ((4, 4), (6, 5)),  # Movimiento en L hacia abajo y derecha
@@ -39,10 +41,11 @@ class TestKnight(unittest.TestCase):
         ]
         for from_pos, to_pos in valid_moves:
             with self.subTest(from_pos=from_pos, to_pos=to_pos):
-                self.assertTrue(self.knight.knight_valid_position(self.board, from_pos, to_pos))
+                from_row, from_col = from_pos
+                to_row, to_col = to_pos
+                self.assertTrue(self.knight.valid_positions(from_row, from_col, to_row, to_col))
 
-# Testeo de movimientos inválidos
-
+    # Testeo de movimientos inválidos
     def test_invalid_moves(self):
         invalid_moves = [
             ((4, 4), (4, 6)),  # Movimiento horizontal (no válido para un caballo)
@@ -52,8 +55,9 @@ class TestKnight(unittest.TestCase):
         ]
         for from_pos, to_pos in invalid_moves:
             with self.subTest(from_pos=from_pos, to_pos=to_pos):
-                self.assertFalse(self.knight.knight_valid_position(self.board, from_pos, to_pos))
-
+                from_row, from_col = from_pos
+                to_row, to_col = to_pos
+                self.assertFalse(self.knight.valid_positions(from_row, from_col, to_row, to_col))
 
 if __name__ == '__main__':
     unittest.main()